style: refactor all components from HrmStyles except form elements
