+++ plugin-hrm-form/src/styles/HrmStyles.tsx
@@ -16,44 +16,11 @@
 
 /* eslint-disable no-nested-ternary */
 import React from 'react';
-import { ButtonBase, Input, MenuItem, MenuList, Select, Tab, TabProps, Tabs, withStyles } from '@material-ui/core';
-import type { ButtonBaseProps } from '@material-ui/core/ButtonBase';
-import AssignmentInd from '@material-ui/icons/AssignmentIndOutlined';
-import { Button, Icon, styled } from '@twilio/flex-ui';
-import { getBackgroundWithHoverCSS } from '@twilio/flex-ui-core';
-import InfoOutlinedIcon from '@material-ui/icons/InfoOutlined';
-import Tooltip from '@material-ui/core/Tooltip';
+import { Input, MenuItem, MenuList, Select } from '@material-ui/core';
+import { styled } from '@twilio/flex-ui';
 
-import { BottomButtonBarHeight } from '.';
 import { Row, Flex } from './layout';
 import HrmTheme from './HrmTheme';
-
-export const ErrorText = styled('p')`
-  color: ${HrmTheme.colors.errorColor};
-  font-size: 10px;
-  line-height: 1.5;
-`;
-ErrorText.displayName = 'ErrorText';
-
-export const CategorySubtitleSection = styled('div')`
-  display: flex;
-  align-items: center;
-  margin: 6px 0;
-`;
-CategorySubtitleSection.displayName = 'CategorySubtitleSection';
-
-export const CategoryRequiredText = styled('p')`
-  color: ${HrmTheme.colors.darkTextColor};
-  font-size: 12px;
-  font-weight: 400;
-  flex-grow: 1;
-
-  &:before {
-    color: ${HrmTheme.colors.errorColor};
-    content: '* ';
-  }
-`;
-CategoryRequiredText.displayName = 'CategoryRequiredText';
 
 const StyledInput = styled(Input)`
   display: flex;
@@ -159,273 +126,6 @@ export const StyledMenuItem = styled(MenuItem)`
   text-transform: capitalize;
 `;
 StyledMenuItem.displayName = 'StyledMenuItem';
-
-type ToggleViewButtonProps = { active?: boolean };
-
-export const ToggleViewButton = styled('button')<ToggleViewButtonProps>`
-  display: inline-flex;
-  width: 37px;
-  height: 37px;
-  min-height: 37px;
-  border-radius: 1px;
-  align-items: center;
-  justify-content: center;
-  box-sizing: border-box;
-  cursor: pointer;
-  border: ${({ active }) => (active ? '1px solid #a0a8bd33' : 'none')};
-  color: ${({ active }) => (active ? '#000000cc' : 'initial')};
-  background-color: ${({ active }) => (active ? 'initial' : '#a0a8bdcc')};
-  opacity: ${({ active }) => (active ? 'initial' : '20%')};
-
-  &:focus {
-    outline: auto;
-  }
-
-  > svg {
-    font-size: 18px;
-  }
-`;
-ToggleViewButton.displayName = 'ToggleViewButton';
-
-export const InformationIconButton = withStyles({
-  root: {
-    width: '16px',
-    height: '16px',
-    color: '#b4babd',
-    margin: '13px 20px 0 5px',
-    cursor: 'pointer',
-  },
-})(InfoOutlinedIcon);
-
-export const HtmlTooltip = withStyles(theme => ({
-  tooltip: {
-    backgroundColor: '#717171',
-    color: '#fff',
-    maxWidth: 400,
-    fontSize: '10pt',
-    fontStyle: 'open sans semibold',
-    border: '1px solid #dadde9',
-  },
-}))(Tooltip);
-
-export const CategoriesWrapper = styled('div')`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  box-sizing: border-box;
-  padding-bottom: ${BottomButtonBarHeight}px;
-`;
-CategoriesWrapper.displayName = 'CategoriesWrapper';
-
-type SubcategoriesWrapperProps = { gridView?: boolean };
-
-export const SubcategoriesWrapper = styled('div')<SubcategoriesWrapperProps>`
-  display: flex;
-  padding: 10px 0 10px 6px;
-  flex-wrap: wrap;
-  flex-direction: ${({ gridView }) => (gridView ? 'row' : 'column')};
-`;
-SubcategoriesWrapper.displayName = 'SubcategoriesWrapper';
-
-export const StyledTabs = withStyles({
-  root: {
-    minHeight: 35,
-    height: 35,
-    flexShrink: 0,
-    padding: '0 7%',
-  },
-  indicator: {
-    backgroundColor: 'transparent',
-  },
-})(Tabs);
-StyledTabs.displayName = 'StyledTabs';
-
-export type StyledTabProps = { searchTab?: boolean; label: React.ReactNode } & Partial<TabProps>;
-
-export const StyledTab = withStyles({
-  root: {
-    height: 35,
-    minHeight: 35,
-    minWidth: 120,
-    width: 120,
-    backgroundColor: '#ecedf1',
-    borderTopLeftRadius: 4,
-    borderTopRightRadius: 4,
-    margin: '0 5px 0 0',
-    padding: 0,
-    fontSize: 12,
-    lineHeight: 14,
-    textTransform: 'none',
-    '&:hover': {
-      backgroundColor: '#c9c9c9',
-    },
-    '&:focus': {
-      outline: 'auto',
-    },
-  },
-  selected: {
-    backgroundColor: '#ffffff',
-    fontWeight: 600,
-  },
-})(Tab);
-StyledTab.displayName = 'StyledTab';
-
-export const StyledSearchTab = withStyles({
-  root: {
-    minWidth: 40,
-    width: 40,
-    backgroundColor: 'transparent',
-    '&:focus': {
-      outline: 'auto',
-    },
-  },
-  selected: {
-    backgroundColor: '#ffffff',
-  },
-})(StyledTab);
-StyledSearchTab.displayName = 'StyledSearchTab';
-
-export const FontOpenSans = styled('p')`
-  color: #000000;
-  font-family: Open Sans;
-  text-align: left;
-`;
-FontOpenSans.displayName = 'FontOpenSans';
-
-// used for screen readers only
-export const HiddenText = styled('span')`
-  width: 0px;
-  height: 0px;
-  font-size: 0px;
-  line-height: 0px;
-`;
-HiddenText.displayName = 'HiddenText';
-
-// This text will not be a child element but can be used for aria-live
-// where the text is read when it enters the dom
-export const AriaLiveHiddenText = styled('span')`
-  top: 0;
-  left: -2px;
-  width: 1px;
-  height: 1px;
-  position: absolute;
-  overflow: hidden;
-`;
-AriaLiveHiddenText.displayName = 'AriaLiveHiddenText';
-
-export const CasePrintViewSpinner = styled('div')`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 100%;
-`;
-CasePrintViewSpinner.displayName = 'CasePrintViewSpinner';
-
-const TaskButtonBase = withStyles({
-  root: {
-    '&:hover': {
-      backgroundColor: '#ECEDF1',
-    },
-    '&:hover > div': {
-      backgroundColor: '#ECEDF1',
-    },
-  },
-  disabled: {
-    opacity: 0.3,
-    color: '#192B33',
-    '& svg': {
-      color: '#192B33',
-    },
-    '& p': {
-      color: '#192B33',
-    },
-  },
-})(ButtonBase);
-
-// @ts-ignore
-export const AddTaskIconContainer = styled('div')`
-  display: flex;
-  flex: 0 0 44px;
-  height: 44px;
-  background-color: #ffffff;
-`;
-AddTaskIconContainer.displayName = 'AddTaskIconContainer';
-
-export const AddTaskIcon = styled(Icon)`
-  display: flex;
-  flex: 0 0 auto;
-  margin: auto;
-  color: #000000;
-`;
-AddTaskIcon.displayName = 'AddTaskIcon';
-
-export const AddTaskContent = styled('div')`
-  display: flex;
-  flex: 1 1 auto;
-  overflow: hidden;
-  padding-right: auto;
-  padding-left: 12px;
-`;
-AddTaskContent.displayName = 'AddTaskContent';
-
-export const AddTaskText = styled(FontOpenSans)`
-  color: #0d74d5;
-  font-size: 12px;
-  line-height: 16px;
-  font-weight: 600;
-  margin-right: 5px;
-`;
-AddTaskText.displayName = 'AddTaskText';
-
-export const AddTaskButtonBase = styled(TaskButtonBase)``;
-AddTaskButtonBase.displayName = 'AddTaskButtonBase';
-
-export const OfflineContactTaskIconContainer = styled('div')`
-  display: flex;
-  flex: 0 0 44px;
-  height: 44px;
-`;
-OfflineContactTaskIconContainer.displayName = 'OfflineContactTaskIconContainer';
-
-export const OfflineContactTaskIcon = withStyles({
-  root: {
-    display: 'flex',
-    flex: '0 0 auto',
-    margin: 'auto',
-    color: '#159af8',
-  },
-})(AssignmentInd);
-OfflineContactTaskIcon.displayName = 'OfflineContactTaskIcon';
-
-export const OfflineContactTaskContent = styled('div')`
-  display: flex;
-  flex: 1 1 auto;
-  flex-direction: column;
-  overflow: hidden;
-  padding-right: auto;
-  padding-left: 12px;
-`;
-OfflineContactTaskContent.displayName = 'OfflineContactTaskContent';
-
-export const OfflineContactTaskFirstLine = styled(FontOpenSans)`
-  font-size: 12px;
-  font-weight: 700;
-  color: rgb(34, 34, 34);
-`;
-OfflineContactTaskFirstLine.displayName = 'OfflineContactTaskFirstLine';
-
-export const OfflineContactTaskSecondLine = styled(FontOpenSans)`
-  font-size: 10px;
-  font-weight: 400;
-  color: rgb(34, 34, 34);
-`;
-OfflineContactTaskSecondLine.displayName = 'OfflineContactTaskSecondLine';
-
-// eslint-disable-next-line react/prop-types
-export const OfflineContactTaskButton: React.FC<{ selected: boolean } & ButtonBaseProps> = ({ selected, ...props }) => (
-  <TaskButtonBase style={{ border: selected ? '2px solid rgb(86, 166, 246)' : 'none' }} {...props} />
-);
-OfflineContactTaskButton.displayName = 'OfflineContactTaskButton';
 
 // eslint-disable-next-line import/no-unused-modules
 export const FormItem = styled('div')`
@@ -891,66 +591,3 @@ export const CategoryCheckboxField = styled('div')<BaseCheckboxProps>`
   cursor: ${({ disabled }) => (disabled ? 'initial' : 'pointer')};
 `;
 CategoryCheckboxField.displayName = 'CategoryCheckboxField';
-
-export const TaskCanvasOverride = styled('div')`
-  width: 100%;
-  height: 100%;
-  background-color: ${HrmTheme.colors.base2};
-`;
-
-export const CannedResponsesContainer = styled('div')`
-  margin-bottom: 15px;
-
-  .form {
-    width: 100%;
-    margin-bottom: 20px;
-  }
-
-  .input-label {
-    padding-left: 5px;
-  }
-`;
-
-CannedResponsesContainer.displayName = 'CannedResponsesContainer';
-
-export const Bold = styled('span')`
-  font-weight: 700;
-`;
-
-Bold.displayName = 'Bold';
-
-export const TypingIndicatorText = styled(FontOpenSans)`
-  font-size: 10px;
-  margin: auto 12px;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-`;
-TypingIndicatorText.displayName = 'TypingIndicatorText';
-
-export const HorizontalLine = styled('hr')`
-  border: 0;
-  height: 1px;
-  background-color: #c6cad7;
-  margin: 0;
-`;
-HorizontalLine.displayName = 'HorizontalLine';
-
-// Used in search results and viewing records of an identifier(View Records in Yello Banner) regardless of client profiles
-export const PreviewWrapper = styled(Flex)`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  margin-top: 10px;
-  padding: 5px 20px 10px 20px;
-  width: 100%;
-  max-width: 100%;
-  box-sizing: border-box;
-  background-color: #ffffff;
-  border-color: #e1e3ea;
-  border-style: solid;
-  border-width: 1px;
-  border-radius: 4px;
-`;
-
-PreviewWrapper.displayName = 'PreviewWrapper';
